refactor: modulize calcalateScore and convertScoreToString
